refactor(es-lint, AllVehicles.js): Enabling es-lint rules prev. disabled; moved AllVehiclesOverlay t
