+++ packages/base-map/src/index.story.js
@@ -1,11 +1,11 @@
 import React from "react";
-import { Marker, CircleMarker, Popup, FeatureGroup } from "react-leaflet";
-import L from "leaflet";
+import { Marker, CircleMarker } from "react-leaflet";
 import { action } from "@storybook/addon-actions";
 import { withInfo } from "@storybook/addon-info";
 
 import BaseMap from ".";
 import SelectVehicles from "../mocks/SelectVehicles";
+import AllVehiclesOverlay from "../mocks/AllVehicles";
 
 import "../assets/map.css";
 
@@ -40,20 +40,11 @@ const mapConfig = {
 
 const center = [mapConfig.initLat, mapConfig.initLon];
 
-const vehicleData = [
-  {
-    vehicleID: 1742,
-    routeNumber: "40",
-    latitude: 33.755558,
-    longitude: -84.393716
-  },
-  {
-    vehicleID: 7024,
-    routeNumber: "03",
-    latitude: 33.754399,
-    longitude: -84.37547
-  }
-];
+const sampleMarkers = (
+  <CircleMarker center={center} radius={100} interactive={false}>
+    <Marker position={center} />
+  </CircleMarker>
+);
 
 const samplePopup = (
   <div>
@@ -62,40 +53,6 @@ const samplePopup = (
       Sample <span style={{ color: "purple" }}>popup</span> content.
     </p>
   </div>
-);
-
-const AllVehiclesOverlay = () => {
-  // Render is according to this file:
-  // https://github.com/OpenTransitTools/transit-components/blob/master/lib/vehicles/AllVehicles.js
-  return (
-    <FeatureGroup className="vehicles">
-      {vehicleData.map(vehicle => {
-        const key = vehicle.vehicleID;
-        const position = [vehicle.latitude, vehicle.longitude];
-
-        return (
-          <Marker
-            class="marker"
-            icon={L.divIcon({
-              html: `<span>${vehicle.routeNumber}</span>`
-            })}
-            key={key}
-            position={position}
-          >
-            <Popup>
-              <span>VEH: {key}</span>
-            </Popup>
-          </Marker>
-        );
-      })}
-    </FeatureGroup>
-  );
-};
-
-const sampleMarkers = (
-  <CircleMarker center={center} radius={100} interactive={false}>
-    <Marker position={center} />
-  </CircleMarker>
 );
 
 const onClick = action("onClick");
